move builders to components/test
